--- src/test/java/me/synology/gooseauthapiserver/GooseAuthApiServerApplicationTests.java
@@ -1,13 +1,0 @@
-package me.synology.gooseauthapiserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GooseAuthApiServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
