+++ src/test/java/me/synology/gooseauthapiserver/common/PostgresContainerTest.java
@@ -1,0 +1,28 @@
+package me.synology.gooseauthapiserver.common;
+
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+@Testcontainers
+public abstract class PostgresContainerTest {
+  private static final String POSTGRES_INFO = "gooseauth";
+
+  @Container
+  public static PostgreSQLContainer<?> postgreSQLContainer = new PostgreSQLContainer<>(
+      DockerImageName.parse("postgres:13"))
+      .withDatabaseName(POSTGRES_INFO)
+      .withUsername(POSTGRES_INFO)
+      .withPassword(POSTGRES_INFO);
+
+  @DynamicPropertySource
+  static void postgresqlProperties(DynamicPropertyRegistry registry) {
+    registry.add("spring.datasource.url", postgreSQLContainer::getJdbcUrl);
+    registry.add("spring.datasource.password", postgreSQLContainer::getPassword);
+    registry.add("spring.datasource.username", postgreSQLContainer::getUsername);
+    registry.add("spring.jpa.hibernate.ddl-auto", () -> "create-drop");
+  }
+}
